Added Makefile to project
Just enter "make" into shell and Makefile ensures every .cpp and .h file
in ./src/ is used to make an executable in ./build/
ISSUE: Pretty neat, though the Makefile strangely goes through the .git files.
Why? I do not know, it explicitly only goes through ./src/ and .git
lives in ./ but at least git seems to be left unimpacted, as strange as
it is
RESOLVED: Turns out it was a typo in Makefile. source directory variable
SRC_DIRS was misspelled as SRS_DIRS. Somehow, everything still worked
fine, it looks like make goes through entire directory if given a
variable that doesn't exist. That or SRS_DIRS is a variable that already
exists. Either way,

				 !!!!!
	 DOUBLE CHECK SPELLING BEFORE USING A COMPLEX MAKEFILE
				 !!!!!

This could have been tragic